+++ compiler/core/js_pass_flatten.ml
@@ -51,7 +51,7 @@ let flatten_map =
             } ->
           S.block
             (Ext_list.map args (fun arg -> self.statement self (S.exp arg)))
-        | Exp {expression_desc = Cond (a, b, c); comment} ->
+        | Exp {expression_desc = Cond (a, b, c); comment; source_loc} ->
           {
             statement_desc =
               If
@@ -59,6 +59,7 @@ let flatten_map =
                   [self.statement self (S.exp b)],
                   [self.statement self (S.exp c)] );
             comment;
+            source_loc;
           }
         | Exp
             {
@@ -76,7 +77,7 @@ let flatten_map =
           (* super#statement *)
           (*   (S.block (List.rev_append rest_rev [S.exp (E.assign a  last_one)])) *)
           | _ -> assert false)
-        | Return {expression_desc = Cond (a, b, c); comment} ->
+        | Return {expression_desc = Cond (a, b, c); comment; source_loc} ->
           {
             statement_desc =
               If
@@ -84,6 +85,7 @@ let flatten_map =
                   [self.statement self (S.return_stmt b)],
                   [self.statement self (S.return_stmt c)] );
             comment;
+            source_loc;
           }
         | Return ({expression_desc = Seq _; _} as v) -> (
           let block = Js_analyzer.rev_flatten_seq v in