+++ compiler/core/lam_compile.ml
@@ -25,6 +25,64 @@
 module E = Js_exp_make
 module S = Js_stmt_make
 
+let source_map_loc = Js_source_map.source_loc_of_loc
+
+let with_source_loc loc (exp : J.expression) =
+  match (source_map_loc loc, exp.source_loc) with
+  | Some source_loc, None -> {exp with source_loc = Some source_loc}
+  | _ -> exp
+
+let rec source_loc_of_lam (lam : Lam.t) =
+  match lam with
+  | Lapply {ap_info = {ap_loc}} -> Some ap_loc
+  | Lprim {loc} | Lfunction {loc} -> Some loc
+  | Llet (_, _, arg, body) -> (
+    match source_loc_of_lam arg with
+    | Some _ as loc -> loc
+    | None -> source_loc_of_lam body)
+  | Lletrec (_, body) | Lsequence (_, body) -> source_loc_of_lam body
+  | Lifthenelse (_, then_, _) -> source_loc_of_lam then_
+  | Lstaticcatch (body, _, _) | Ltrywith (body, _, _) -> source_loc_of_lam body
+  | Lstringswitch (_, cases, default) -> (
+    match cases with
+    | (_, body) :: _ -> source_loc_of_lam body
+    | [] -> (
+      match default with
+      | Some body -> source_loc_of_lam body
+      | None -> None))
+  | Lswitch (_, sw) -> (
+    match (sw.sw_consts, sw.sw_blocks, sw.sw_failaction) with
+    | (_, body) :: _, _, _ | _, (_, body) :: _, _ -> source_loc_of_lam body
+    | [], [], Some body -> source_loc_of_lam body
+    | [], [], None -> None)
+  | Lstaticraise (_, args) -> (
+    match args with
+    | arg :: _ -> source_loc_of_lam arg
+    | [] -> None)
+  | Lwhile (_, body)
+  | Lfor (_, _, _, _, body)
+  | Lfor_of (_, _, body)
+  | Lfor_await_of (_, _, body) ->
+    source_loc_of_lam body
+  | Lassign (_, body) -> source_loc_of_lam body
+  | Lvar _ | Lglobal_module _ | Lconst _ | Lbreak | Lcontinue -> None
+
+let source_map_loc_of_lam lam =
+  match source_loc_of_lam lam with
+  | Some loc -> source_map_loc loc
+  | None -> None
+
+let with_statement_source_loc source_loc (stmt : J.statement) =
+  match (source_loc, stmt.source_loc) with
+  | Some source_loc, None -> {stmt with source_loc = Some source_loc}
+  | _ -> stmt
+
+let with_block_source_loc lam block =
+  match block with
+  | [] -> []
+  | stmt :: rest ->
+    with_statement_source_loc (source_map_loc_of_lam lam) stmt :: rest
+
 let args_either_function_or_const (args : Lam.t list) =
   Ext_list.for_all args (fun x ->
       match x with
@@ -315,6 +373,7 @@ let compile output_prefix =
           | Single x ->
             apply_with_arity fn ~arity:(Lam_arity.extract_arity x) args)
       in
+      let expression = with_source_loc appinfo.ap_info.ap_loc expression in
       Js_output.output_of_block_and_expression lambda_cxt.continuation args_code
         expression
   (*
@@ -329,7 +388,12 @@ let compile output_prefix =
       (id : Ident.t) (arg : Lam.t) : Js_output.t * initialization =
     match arg with
     | Lfunction
-        {params; body; attr = {return_unit; async; one_unit_arg; directive}} ->
+        {
+          params;
+          body;
+          attr = {return_unit; async; one_unit_arg; directive};
+          loc;
+        } ->
       (* TODO: Think about recursive value
          {[
            let rec v = ref (fun _ ...
@@ -386,6 +450,7 @@ let compile output_prefix =
             (Js_output.output_as_block output)
             ~return_unit ~async ~one_unit_arg ?directive
       in
+      let result = with_source_loc loc result in
       ( Js_output.output_of_expression
           (Declare (Alias, id))
           result
@@ -539,8 +604,12 @@ let compile output_prefix =
           (a * J.case_clause) list ->
           J.statement) ~(switch_exp : J.expression) ~(default : default_case)
        ?(merge_cases = fun _ _ -> true) (cases : (a * Lam.t) list) ->
+    let output_block_with_source_loc cxt lam =
+      compile_lambda cxt lam |> Js_output.output_as_block
+      |> with_block_source_loc lam
+    in
     match (cases, default) with
-    | [], Default lam -> Js_output.output_as_block (compile_lambda cxt lam)
+    | [], Default lam -> output_block_with_source_loc cxt lam
     | [], (Complete | NonComplete) -> []
     | [(_, lam)], Complete ->
       (* To take advantage of such optimizations,
@@ -549,18 +618,18 @@ let compile output_prefix =
           otherwise the compiler engine would think that
           it's also complete
       *)
-      Js_output.output_as_block (compile_lambda cxt lam)
+      output_block_with_source_loc cxt lam
     | [(id, lam)], NonComplete ->
       morph_declare_to_assign cxt (fun cxt define ->
           [
             S.if_ ?declaration:define
               (eq_exp None switch_exp (Some id) (make_exp id))
-              (Js_output.output_as_block (compile_lambda cxt lam));
+              (output_block_with_source_loc cxt lam);
           ])
     | [(id, lam)], Default x | [(id, lam); (_, x)], Complete ->
       morph_declare_to_assign cxt (fun cxt define ->
-          let else_block = Js_output.output_as_block (compile_lambda cxt x) in
-          let then_block = Js_output.output_as_block (compile_lambda cxt lam) in
+          let else_block = output_block_with_source_loc cxt x in
+          let then_block = output_block_with_source_loc cxt lam in
           [
             S.if_ ?declaration:define
               (eq_exp None switch_exp (Some id) (make_exp id))
@@ -599,9 +668,7 @@ let compile output_prefix =
             | Complete -> None
             | NonComplete -> None
             | Default lam -> (
-              let statements =
-                Js_output.output_as_block (compile_lambda switch_cxt lam)
-              in
+              let statements = output_block_with_source_loc switch_cxt lam in
               match statements with
               | [] -> None
               | _ -> Some statements)
@@ -613,6 +680,7 @@ let compile output_prefix =
                   let switch_body, should_break =
                     Js_output.to_break_block (compile_lambda switch_cxt lam)
                   in
+                  let switch_body = with_block_source_loc lam switch_body in
                   let should_break =
                     if
                       not
@@ -621,10 +689,22 @@ let compile output_prefix =
                     then should_break
                     else should_break && Lam_exit_code.has_exit lam
                   in
-                  (switch_case, J.{switch_body; should_break; comment = None})
+                  ( switch_case,
+                    J.
+                      {
+                        switch_body;
+                        should_break;
+                        comment = None;
+                        source_loc = source_map_loc_of_lam lam;
+                      } )
                 else
                   ( switch_case,
-                    {switch_body = []; should_break = false; comment = None} ))
+                    {
+                      switch_body = [];
+                      should_break = false;
+                      comment = None;
+                      source_loc = source_map_loc_of_lam lam;
+                    } ))
             (* TODO: we should also group default *)
             (* The last clause does not need [break]
                 common break through, *)
@@ -1630,11 +1710,12 @@ let compile output_prefix =
       | _ ->
         Js_output.output_of_block_and_expression lambda_cxt.continuation
           args_code
-          (E.call
-             ~info:
-               (call_info_of_ap_status appinfo.ap_transformed_jsx
-                  appinfo.ap_info.ap_status)
-             fn_code args))
+          (with_source_loc appinfo.ap_info.ap_loc
+             (E.call
+                ~info:
+                  (call_info_of_ap_status appinfo.ap_transformed_jsx
+                     appinfo.ap_info.ap_status)
+                fn_code args)))
   and compile_prim (prim_info : Lam.prim_info)
       (lambda_cxt : Lam_compile_context.t) =
     match prim_info with
@@ -1648,13 +1729,16 @@ let compile output_prefix =
       | Fld_module {name = field} ->
         compile_external_field ~dynamic_import lambda_cxt id field
       | _ -> assert false)
-    | {primitive = Praise; args = [e]; _} -> (
+    | {primitive = Praise; args = [e]; loc} -> (
       match
         compile_lambda {lambda_cxt with continuation = NeedValue Not_tail} e
       with
       | {block; value = Some v} ->
+        let stmt =
+          with_statement_source_loc (source_map_loc loc) (S.throw_stmt v)
+        in
         Js_output.make
-          (Ext_list.append_one block (S.throw_stmt v))
+          (Ext_list.append_one block stmt)
           ~value:E.undefined ~output_finished:True
       (* FIXME -- breaks invariant when NeedValue, reason is that js [throw] is statement
          while ocaml it's an expression, we should remove such things in lambda optimizations
@@ -1662,11 +1746,14 @@ let compile output_prefix =
       | {value = None} -> assert false)
     | {primitive = Psequand; args = [l; r]; _} -> compile_sequand l r lambda_cxt
     | {primitive = Psequor; args = [l; r]} -> compile_sequor l r lambda_cxt
-    | {primitive = Pdebugger; _} ->
+    | {primitive = Pdebugger; loc; _} ->
       (* [%debugger] guarantees that the expression does not matter
          TODO: make it even safer *)
-      Js_output.output_of_block_and_expression lambda_cxt.continuation
-        S.debugger_block E.unit
+      let stmt =
+        with_statement_source_loc (source_map_loc loc) (S.debugger_stmt ())
+      in
+      Js_output.output_of_block_and_expression lambda_cxt.continuation [stmt]
+        E.unit
       (* TODO:
          check the arity of fn before wrapping it
          we need mark something that such eta-conversion can not be simplified in some cases
@@ -1721,24 +1808,25 @@ let compile output_prefix =
     | {primitive = Pjs_unsafe_downgrade _; args} -> assert false
     | {primitive = Pjs_fn_method; args = args_lambda} -> (
       match args_lambda with
-      | [Lfunction {params; body; attr = {return_unit; async}}] ->
+      | [Lfunction {params; body; attr = {return_unit; async}; loc}] ->
         Js_output.output_of_block_and_expression lambda_cxt.continuation []
-          (E.method_ ~async ~return_unit params
-             (* Invariant:  jmp_table can not across function boundary,
-                here we share env
-             *)
-             (Js_output.output_as_block
-                (compile_lambda
-                   {
-                     lambda_cxt with
-                     continuation =
-                       EffectCall
-                         (Maybe_tail_is_return
-                            (Tail_with_name
-                               {label = None; in_staticcatch = false}));
-                     jmp_table = Lam_compile_context.empty_handler_map;
-                   }
-                   body)))
+          (with_source_loc loc
+             (E.method_ ~async ~return_unit params
+                (* Invariant:  jmp_table can not across function boundary,
+                   here we share env
+                *)
+                (Js_output.output_as_block
+                   (compile_lambda
+                      {
+                        lambda_cxt with
+                        continuation =
+                          EffectCall
+                            (Maybe_tail_is_return
+                               (Tail_with_name
+                                  {label = None; in_staticcatch = false}));
+                        jmp_table = Lam_compile_context.empty_handler_map;
+                      }
+                      body))))
       | _ -> assert false)
     | {primitive = Pjs_fn_make arity; args = [fn]; loc} ->
       compile_lambda lambda_cxt
@@ -1782,7 +1870,7 @@ let compile output_prefix =
             [args_expr]
         in
         Js_output.output_of_block_and_expression lambda_cxt.continuation
-          args_code exp
+          args_code (with_source_loc loc exp)
       | Lfunction
           {
             body =
@@ -1811,7 +1899,7 @@ let compile output_prefix =
             [args_expr]
         in
         Js_output.output_of_block_and_expression lambda_cxt.continuation
-          args_code exp
+          args_code (with_source_loc loc exp)
       | _ ->
         Location.raise_errorf ~loc
           "Invalid argument: unsupported argument to dynamic import. If you \
@@ -1833,7 +1921,7 @@ let compile output_prefix =
           args_expr
       in
       Js_output.output_of_block_and_expression lambda_cxt.continuation args_code
-        exp
+        (with_source_loc loc exp)
   and collect_dup_overrides (copy_id : Ident.t) (lam : Lam.t)
       (acc : (Lam_compat.set_field_dbg_info * Lam.t) list) :
       (Lam_compat.set_field_dbg_info * Lam.t) list option =
@@ -1849,7 +1937,7 @@ let compile output_prefix =
   and try_compile_record_spread (lambda_cxt : Lam_compile_context.t)
       (id : Ident.t) (arg : Lam.t) (body : Lam.t) : Js_output.t option =
     match arg with
-    | Lprim {primitive = Pduprecord; args = [init]; _} -> (
+    | Lprim {primitive = Pduprecord; args = [init]; loc} -> (
       match collect_dup_overrides id body [] with
       | None -> None
       | Some overrides ->
@@ -1885,32 +1973,39 @@ let compile output_prefix =
         Some
           (Js_output.output_of_block_and_expression lambda_cxt.continuation
              blocks
-             (E.obj ~dup:init_val props)))
+             (with_source_loc loc (E.obj ~dup:init_val props))))
     | _ -> None
   and compile_lambda (lambda_cxt : Lam_compile_context.t) (cur_lam : Lam.t) :
       Js_output.t =
     match cur_lam with
     | Lfunction
-        {params; body; attr = {return_unit; async; one_unit_arg; directive}} ->
+        {
+          params;
+          body;
+          attr = {return_unit; async; one_unit_arg; directive};
+          loc;
+        } ->
       Js_output.output_of_expression lambda_cxt.continuation
         ~no_effects:no_effects_const
-        (E.ocaml_fun params ~return_unit ~async ~one_unit_arg ?directive
-           (* Invariant:  jmp_table can not across function boundary,
-              here we share env
-           *)
-           (Js_output.output_as_block
-              (compile_lambda
-                 {
-                   lambda_cxt with
-                   continuation =
-                     EffectCall
-                       (Maybe_tail_is_return
-                          (Tail_with_name {label = None; in_staticcatch = false}));
-                   jmp_table = Lam_compile_context.empty_handler_map;
-                   switch_depth = 0;
-                   loop_stack = [];
-                 }
-                 body)))
+        (with_source_loc loc
+           (E.ocaml_fun params ~return_unit ~async ~one_unit_arg ?directive
+              (* Invariant:  jmp_table can not across function boundary,
+                 here we share env
+              *)
+              (Js_output.output_as_block
+                 (compile_lambda
+                    {
+                      lambda_cxt with
+                      continuation =
+                        EffectCall
+                          (Maybe_tail_is_return
+                             (Tail_with_name
+                                {label = None; in_staticcatch = false}));
+                      jmp_table = Lam_compile_context.empty_handler_map;
+                      switch_depth = 0;
+                      loop_stack = [];
+                    }
+                    body))))
     | Lapply appinfo -> compile_apply appinfo lambda_cxt
     | Llet (let_kind, id, arg, body) -> (
       match try_compile_record_spread lambda_cxt id arg body with